requirements.txt
connexion==2.14.2
swagger-ui-bundle==0.0.9
Flask-JWT-Extended==4.5.2
flask_sqlalchemy==3.0.5
bcrypt==4.0.1
pandas==2.0.2